do not crash in case of duplicated date+slug
It is technically possible for the user to create two entries on the same
day that have the same slug.  This choice will however make it problematic
to distinguish between the two entries.

Even if this is not a sane situation, PFT shoudl not crash.  It will
instead cast a warning.